rename: 컴포넌트,인터페이스 네이밍 변경, 검증 로직 메세지 상수화
* 컴포넌트, 인터페이스 네이밍 변경
  - styled-components로 정의한 스타일 컴포넌트의 LI, UL 네이밍을 Li, Ul로 변경
  - iTodo 인터페이스를 ITodo로 변경

* 검증 로직 메세지 상수화
  - 이메일, 비밀번호 검증 로직 메세지 상수화해서 외부 utils 폴더에 빼내고 외부에서 공급받고
  외부에서 변경가능하게 수정함

2023.08.07
