+++ src/pages/Todo.tsx
@@ -4,13 +4,13 @@ import TodoList from '../components/TodoList'
 import { useNavigate } from 'react-router-dom'
 import AppHeader from '../components/AppHeader'
 import { createTodo, deleteTodo, getTodos, updateTodo } from '../api/TodoApi'
-import { TodoWrapper, TodoMain, TodoTitle, UL } from '../styles/TodoStyle'
+import { TodoWrapper, TodoMain, TodoTitle, Ul } from '../styles/TodoStyle'
 import useError from '../Hooks/useError'
 import localToken from '../api/LocalToken'
 
 export type onAddFunction = (todo: string) => void
-export type OnTodoFunction = (todoItem: iTodo) => void
-export interface iTodo {
+export type OnTodoFunction = (todoItem: ITodo) => void
+export interface ITodo {
   id: number
   todo: string
   isCompleted: boolean
@@ -18,7 +18,7 @@ export interface iTodo {
 }
 
 export default function Todo() {
-  const [todos, setTodos] = useState<iTodo[]>([])
+  const [todos, setTodos] = useState<ITodo[]>([])
   const { showError } = useError()
 
   const accessToken = localToken.get()
@@ -50,7 +50,7 @@ export default function Todo() {
     }
   }
 
-  const handleCheck = (todoItem: iTodo) => {
+  const handleCheck = (todoItem: ITodo) => {
     try {
       updateTodo(todoItem)
       setTodos(todos.map((item) => (item.id === todoItem.id ? todoItem : item)))
@@ -59,7 +59,7 @@ export default function Todo() {
     }
   }
 
-  const handleEdit = (todoItem: iTodo) => {
+  const handleEdit = (todoItem: ITodo) => {
     try {
       updateTodo(todoItem)
       setTodos(
@@ -93,7 +93,7 @@ export default function Todo() {
         <TodoTitle>TODO LIST</TodoTitle>
         <TodoInput onAdd={handleAdd} />
         {todos?.length > 0 && (
-          <UL className="todos">
+          <Ul className="todos">
             {todos.map((item) => (
               <TodoList
                 key={item.id}
@@ -103,14 +103,14 @@ export default function Todo() {
                 onDelete={handleDelete}
               />
             ))}
-          </UL>
+          </Ul>
         )}
       </TodoMain>
     </TodoWrapper>
   )
 }
 
-async function readTodo(): Promise<iTodo[]> {
+async function readTodo(): Promise<ITodo[]> {
   const res = await getTodos()
   return res
 }
