+++ src/utils/message.ts
@@ -1,0 +1,4 @@
+export const UNKNOWN_ERROR = '알 수 없는 에러가 발생했습니다.'
+
+export const VALID_MESSAGE_EMAIL = '이메일에 @ 기호가 없습니다.'
+export const VALID_MESSAGE_PASSWORD = '비밀번호는 8자 이상이여야 합니다.'
