--- src/utils/unknownError.ts
@@ -1,1 +1,0 @@
-export const UNKNOWN_ERROR = '알 수 없는 에러가 발생했습니다.'
